\name{ Fitting functions }
\alias{fun.fit.gl.v2a.nw}
\alias{fun.fit.gl.v2a}
\alias{fun.fit.gl.v2b.nw}
\alias{fun.fit.gl.v2b}
\alias{fun.fit.gl.v3}
\alias{fun.fit.gl.v3a}
\alias{fun.fit.gl.v3m}
\alias{fun.fit.gl.v4}
\alias{fun.fit.gl.v4a}
\alias{fun.fit.gl.v6}
\alias{fun.fit.gl.v6a}
\alias{fun.fit.gl.v.lm}
\alias{fun.fit.gl.v.lma}


%- Also NEED an '\alias' for EACH other topic documented here.
\title{ This is a collection of functions designed to implement the fitting 
algorithms for all the methods covered
in this package. }
\description{
These functions are for maintainers only and it is not designed for the users 
of this package.
}
\author{ Steve Su }
\value{ List of unique optimisation parameters and objective values under different fitting methods, not intended to be used directly by users of this package }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal}