\name{ FMKL fitting and basic functions }
\alias{fun.auto.mm.fmkl}
\alias{fun.fmkl.mm.min}
\alias{fun.fmkl.mm.sol.alt}
\alias{fun.fmkl.mm.sol}
\alias{fun.fmkl.sol}
\alias{fun.fmkl.nr}
\alias{fun.fmkl}
\alias{fun.fmkla}
\alias{fun.fmklb}
\alias{fun.fmkl0}
\alias{fun.fmkl.L30}
\alias{fun.fmkl.L40}
\alias{fun.Lm.gt.2.fmkl}
\alias{fun.moments}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ This is a collection of functions designed to find the initial values 
under the method of moments for FMKL
generalised lambda distribution. It also includes basic FMKL GLD functions. }
\description{
These functions are for maintainers only and it is not designed for the users 
of this package.
}
\author{ Steve Su }
\value{ List of computation results for various fitting algorithms, not intended to be used by users of this package }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal}