\name{fun.beta}
\alias{fun.beta}
\alias{fun.beta1}
\title{ This is a collection of functions used in the calculation of the beta 
function. }
\description{
These functions are for maintainers only and it is not designed for the users of 
this package.
}
\author{ Steve Su }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal}

