\name{ Hidden basic functions }
\alias{.gl.parameter.tidy}
\alias{.qdgl.fmkl}
\alias{.qdgl.rs}
\alias{.qgl.rs}
\alias{.qgl.fmkl}
\alias{.qgl.fmkl}
\alias{C_check_gld}
\alias{C_dgl}
\alias{C_mult_check_gld}
\alias{C_optim_fun3}
\alias{C_optim_fun3_v}
\alias{C_pgl} 
\alias{C_q_fmkl_gld_minmax_check}
\alias{C_q_rs_gld_minmax_check}


%- Also NEED an \alias for EACH other topic documented here.
\title{ This is a collection of functions designed to implement the basic GLD 
functions. }
\description{
These functions are for maintainers only and it is not designed for the users of 
this package.
}
\author{ Steve Su }
\value{ Various basic calculations for generalised lambda distributions, not intended to be used by users of this package }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal} 