\name{Optimisation functions}
\alias{optim.fun2.nw}
\alias{optim.fun2}
\alias{optim.fun3}
\alias{optim.fun4}
\alias{optim.fun5}
\alias{optim.fun6}
\alias{optim.fun7}
\alias{optim.fun.lm}
\alias{optim.fun.qs}
\alias{optim.fun.bi.final}
\title{ This is a collection of functions used in the optimisation processes for 
all the fitting
methods covered in this package. }
\description{
These functions are for maintainers only and it is not designed for the users of 
this package.
}
\author{ Steve Su }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal} 