\name{ Hidden basic functions }
\alias{.gl.parameter.tidy}
\alias{.qdgl.fmkl}
\alias{.qdgl.rs}
\alias{.qgl.rs}
\alias{.qgl.fmkl}
\alias{.qgl.fmkl}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{ This is a collection of functions designed to implement the basic RS GLD 
functions. }
\description{
These functions are for maintainers only and it is not designed for the users of 
this package.
}
\author{ Steve Su }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal} 