\name{ RS fitting and basic functions }
\alias{fun.auto.perc.rs}
\alias{fun.lambda.percentile}
\alias{fun.percentile}
\alias{fun.rs.nr}
\alias{fun.rs.perc.gradient}
\alias{fun.rs.perc.min}
\alias{fun.rs.perc.sol.alt}
\alias{fun.rs.perc.sol}
\alias{fun.rsb}
\alias{fun.Lm.gt.2.rs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ This is a collection of functions designed to find the initial values of 
method of moments for RS
generalised lambda distribution. It also includes basic RS GLD functions. }
\description{
These functions are for maintainers only and it is not designed for the users of 
this package.
}
\author{ Steve Su }
\note{ Please contact the author directly if you find a bug! }
\keyword{internal} 